+++ mission001/jisun/index.html
@@ -62,6 +62,6 @@
         </ul>
       </div>
     </section>
-    <script src="js/app.js"></script>
+    <script src="js/app.js" type="module"></script>
   </body>
 </html>
