+++ mission001/jisun/js/components/TodoCount.js
@@ -1,0 +1,9 @@
+const todoCount = document.getElementById("todoCount");
+
+export function TodoCount (data) {
+  renderCount(data, todoCount);
+}
+
+export const renderCount = (data, selector) => {
+  selector.innerHTML = `총 <strong>${data.length}</strong> 개`;
+}
